Removes "Visualizations" tab and references to it (#808)
The entire "Visualizations" header tab is devoted to the now-broken Data Studio
dashboards that Chris created a long time ago. Those dashboards relied on the
stats-pipeline, but the stats-pipeline was shut down as a cost saving measure
and because nothing was really using it, other than those dashboards.
Additionally, the dashboard were unmaintained and very slow and possibly broken
in other ways.