Updates links to Mobiperf archival data in GCS (#816)
It used to be in some bucket that we didn't manage, and the links were broken.
I worked with Dave Choffnes to get at the data, but a random user found a way
to download all the data and shared it with me. I downloaded the data and
uploaded it to archive-measurement-lab/mobiperf.