Configures status map to use Locate registration information
The old siteinfo geo format does not include host-managed servers. The
information from Locate is not as rich, but is probably good enough for the
map. Additionally Locate's database doesn't know about "sites", but instead
knows about heartbeat messages from individual experiments, which has the
effect of breaking out M-Lab traditional site markes into mlab1, mlab2, mlab3.
This may not be a bad thing, since it gives a better picture of which machines
are out there, not just which sites.