Changes "hostname" property for "machine" in status map
Full M-Lab hostnames are too long. The "machine" will consist of
"<site>-<machine>". This is reversed from historical M-lab designations which
are "<machine>-<site>", but this is how autojoined machines are named, and so
we go with that because it would be awkward to have "<hex IP addr>-<site>" for
autojoined machines.