Update infrastructure map to use data from Locate instead of siteinfo (#835)
* Disables caching for GCS objects in website buckets

Only for sandbox and staging, since I believe the production site is served out
of Firebase, not with a GCS bucket. This will make it easier to test changes in
sandbox and staging, not having to wait up to an hour for the cache on an
object to expire.

* Changes double quotes to single quotes in Travis build file

I mistakenly had nested doubled quotes, breaking the config.

* Removes the setmeta keywork from gsutil command in Travis config

* Updates status map to use new Locate registrations geo format

This will display not only M-Lab-managed machines, but also BYOS machines. The
data is not as rich as what siteinfo has, but is sufficient for a basic status
map.

* Adds missing semicolon to infrastructure-map.js

* Moves misplaced </h4> from code into an actual string

* Removes stray "</h4>" at the head of the javascript code

I must have inserted that accidentally, and not noticed it when inspecting
git-diff.

* Limits infrastructure map to NDT

Without this additional filter, we will get "nodes" for every experiment in
every location, which gives a distorted view of the platform from a machine
perspective.

* Removes "<br>" after machine name for infrastructure map

* Operate on e.features[0], not aFeatures[c] on infrastructure map

I was mistakenly operating on the wrong object for unclustered points.

* Updates language on the infrastructure status map page

The previous language largely addressed typical M-Lab physical sites. The new
language makes some improvements, and also includes some basic information
about host-managed servers.